:recycle: refactor: Fix Pull Request comments
