+++ static/js/modules/forms/utils.js
@@ -30,3 +30,18 @@ export function formDataToJSON(formData) {
   formData.forEach((val, key) => obj[key] = val.trim());
   return JSON.stringify(obj);
 }
+
+export function getCookie(name) {
+    let cookieValue = null;
+    if (document.cookie && document.cookie !== "") {
+        const cookies = document.cookie.split(";");
+        for (let cookie of cookies) {
+            cookie = cookie.trim();
+            if (cookie.startsWith(name + "=")) {
+                cookieValue = decodeURIComponent(cookie.substring(name.length + 1));
+                break;
+            }
+        }
+    }
+    return cookieValue;
+}
